chore(pkger): refactor parser notification endpoint state out into stateful type
references: #17434
